+++ extensions/tuya-smart/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Tuya Smart Changelog
 
 ## [Initial Version] - 2023-02-09
+
+- Added Tuya Smart Command
+- Added Support for Switches
+- Added Support for Courtains
+- Added Support for Sockets